tests/stream-properties.html

<!DOCTYPE html>
<script src="resources/text-encode-transform.js"></script>
<script src="resources/transform-stream.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

function assertGetterConfiguredCorrectly(prototype, name) {
  const descriptor = Object.getOwnPropertyDescriptor(prototype, name);
  assert_true(descriptor.configurable, 'getter must be configurable');
  assert_true(descriptor.enumerable, 'getter must be enumerable');
  assert_not_equals(descriptor.get, undefined, 'get must be defined');
  assert_equals(descriptor.set, undefined, 'set must be undefined');
}

test(() => {
  const te = new TextEncoder();
  assert_inherits(te, 'readable', 'readable property must be present');
  assertGetterConfiguredCorrectly(TextEncoder.prototype, 'readable');
  assert_readonly(te, 'readable', 'readable property must be read-only');
  assert_equals(typeof ReadableStream.prototype.getReader.call(te.readable),
                'object', 'readable property must pass brand check');
  assert_inherits(te, 'writable', 'writable property must be present');
  assertGetterConfiguredCorrectly(TextEncoder.prototype, 'writable');
  assert_readonly(te, 'writable', 'writable property must be read-only');
  assert_equals(typeof WritableStream.prototype.getWriter.call(te.writable),
                'object', 'writable property must pass brand check');
}, 'TextEncoder object must have readable and writable properties');

test(() => {
  const td = new TextDecoder();
  assert_inherits(td, 'readable', 'readable property must be present');
  assertGetterConfiguredCorrectly(TextDecoder.prototype, 'readable');
  assert_readonly(td, 'readable', 'readable property must be read-only');
  assert_equals(typeof ReadableStream.prototype.getReader.call(td.readable),
                'object', 'readable property must pass brand check');
  assert_inherits(td, 'writable', 'writable property must be present');
  assertGetterConfiguredCorrectly(TextDecoder.prototype, 'writable');
  assert_readonly(td, 'writable', 'writable property must be read-only');
  assert_equals(typeof WritableStream.prototype.getWriter.call(td.writable),
                'object', 'writable property must pass brand check');
}, 'TextDecoder object must have readable and writable properties');

</script>
